WMTS Capabilities in extending/inheriting class
