+++ src/Generic/ServiceCapabilities.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Nieuwland\OgcSerializer\Generic;
 
 use InvalidArgumentException;
-use Nieuwland\OgcSerializer\Type\WMTS\Capabilities\v10\TileMatrixSet;
 
 /**
  * Holds some common used props for OGC services.
@@ -19,23 +18,16 @@ class ServiceCapabilities implements ServiceCapabilitiesInterface
     private $layers = [];
     /** @var string[] */
     private $versions = [];
-    /** @var TileMatrixSet[]|null */
-    private $tileMatrixSets = null;
 
     /**
      * @param LayerCapabilitiesInterface[] $layers
      * @param string[]                     $versions
-     * @param TileMatrixSet[]|null         $tileMatrixSets
      */
-    public function __construct(string $title, array $layers, array $versions, array $tileMatrixSets = null)
+    public function __construct(string $title, array $layers, array $versions)
     {
         $this->title    = $title;
         $this->versions = $versions;
         $this->setLayers($layers);
-        if (null === $tileMatrixSets) {
-            return;
-        }
-        $this->setTileMatrixSets($tileMatrixSets);
     }
 
     public function getTitle(): string
@@ -86,27 +78,5 @@ class ServiceCapabilities implements ServiceCapabilitiesInterface
     public function getVersions(): array
     {
         return $this->versions;
-    }
-
-    public function hasTileMatrixSet(string $identifier): bool
-    {
-        return null !== $this->tileMatrixSets && isset($this->tileMatrixSets[$identifier]);
-    }
-
-    public function getTileMatrixSet(string $identifier): TileMatrixSet
-    {
-        if (false === $this->hasTileMatrixSet($identifier)) {
-            throw new InvalidArgumentException('unknown tile matrix set id');
-        }
-
-        return $this->tileMatrixSets[$identifier];
-    }
-
-    public function setTileMatrixSets(array $tileMatrixSets): void
-    {
-        $this->tileMatrixSets = [];
-        foreach ($tileMatrixSets as $tileMatrixSet) {
-            $this->tileMatrixSets[$tileMatrixSet->getIdentifier()] = $tileMatrixSet;
-        }
     }
 }
