+++ src/Generic/ServiceCapabilitiesInterface.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Nieuwland\OgcSerializer\Generic;
 
-use Nieuwland\OgcSerializer\Type\WMTS\Capabilities\v10\TileMatrixSet;
-
 /**
  * Data Transfer Object of transformed service capabilities for (javascript) clients.
  */
@@ -26,10 +24,4 @@ interface ServiceCapabilitiesInterface
      * @return string[]
      */
     public function getVersions(): array;
-
-    public function hasTileMatrixSet(string $identifier): bool;
-
-    public function getTileMatrixSet(string $identifier): TileMatrixSet;
-
-    public function setTileMatrixSets(array $tileMatrixSets): void;
 }
